requirements.txt
django
djangorestframework
pytest
pytest-django
python-dotenv
flake8
black
django-mptt